backend/requirements.txt
# ── Core ───────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
pydantic==2.7.1
python-dotenv==1.0.1

# ── Groq LLM ───────────────────────────────────────────────────────────────────
groq>=0.9.0

# ── Local embeddings for Pinecone ──────────────────────────────────────────────
sentence-transformers>=2.7.0

# ── Pinecone ───────────────────────────────────────────────────────────────────
pinecone-client>=3.0.0

# ── Gmail API ──────────────────────────────────────────────────────────────────
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.128.0
